Add files via upload
This project aims to build an automatic smart vending machine that controls the
transactions of books, fines etc in a library without the user needing to touch the
surface of the machine or come in physical contact with any other human.
The motivation behind this project is the rising challenges caused due to the
most recent Covid-19 Pandemic. The most common method by which this virus
gets transmitted is through physical touch or transfer from an inanimate object
that is touched by an infected person.
Thus, for the normal functioning of society we will need devices that can be
controlled without touch.
One of the best examples of transactions that happen in colleges are the
borrowing,exchanging and returning of library books. Hence, this project aims
at safe transactions in a library through a smart contactless vending machine
system.